Update TextRepeaterButton.js
Using useState hook from React to store state and set state for the button.

Created an array that will be responsible for storing the number of times "I like this text" is repeated and rendered as a span element.

Note since we will be displaying the array in React, a key value is required and hence why it's placed in the html tag for <span>